refactor: split service utilities
